backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0

# Azure services
azure-storage-blob==12.19.0
azure-ai-documentintelligence==1.0.0b1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# AWS services (legacy - remove after migration)
boto3==1.34.0
botocore==1.34.0

# AI and ML
anthropic>=0.7.0
openai>=1.3.0
langchain>=0.0.300
langchain-community
langchain-text-splitters
pgvector>=0.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==3.17.1
tika==2.6.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1
kombu==5.3.4
celery[sqs]==5.3.4

# Billing
stripe==7.8.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3

# Security
redis==5.0.1

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-boto3sqs==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0
structlog==23.2.0
python-json-logger==2.0.7
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1